requirements.txt
langchain
langchain-community
streamlit
sentence-transformers
faiss-cpu
beautifulsoup4
pyyaml
ragas
datasets
ollama